emojicache: prevent error message spam
this commit fixes a infinite loop that can occur, if there is a message
that can't be decoded. Until now the bot would try to decode the message
indefinitely and keep printing Errors.
Now the bot will instead stop counting the messages but keep the lock
for now since it will most likely be only be able to decode the message,
if it is updated which includes a restart which releases the lock